pyproject.toml
[project]
name = "job-research-tool"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "tavily-python>=0.3.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
]

[project.scripts]
headhunter = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
